reorg: rename game-data and organize factories
