+++ src/js/factories/monster-factory.js
